+++ components/clientComponents/forms/Form/Form.tsx
@@ -2,8 +2,9 @@
 import React, { useEffect, useState, useCallback } from "react";
 import { withFormik } from "formik";
 import { getFormInitialValues } from "@lib/formBuilder";
-import { getErrorList, setFocusOnErrorMessage, validateOnSubmit } from "@lib/validation/validation";
+import { getErrorList, setFocusOnErrorMessage } from "@lib/validation/validation";
 import { Alert, RichText } from "@clientComponents/forms";
+import { validateOnSubmit } from "@gcforms/core";
 
 import { type FormProps, type InnerFormProps } from "./types";
 import { type Language } from "@lib/types/form-builder-types";
@@ -40,6 +41,7 @@ import { CaptchaFail } from "@clientComponents/globals/FormCaptcha/CaptchaFail";
 import { ga } from "@lib/client/clientHelpers";
 import { SubmitProgress } from "@clientComponents/forms/SubmitProgress/SubmitProgress";
 import { handleUploadError } from "@lib/fileInput/handleUploadError";
+import { hasFiles } from "@lib/fileExtractor";
 
 import {
   copyObjectExcludingFileContent,
@@ -145,7 +147,7 @@ const InnerForm: React.FC<InnerFormProps> = (props) => {
   return status === "submitting" ? (
     <>
       <title>{t("loading")}</title>
-      <SubmitProgress />
+      <SubmitProgress spinner={!hasFiles(props.values)} />
     </>
   ) : (
     <>
@@ -319,12 +321,35 @@ export const Form = withFormik<FormProps, Responses>({
       const { formValuesWithoutFileContent, fileObjsRef } =
         copyObjectExcludingFileContent(formValues);
 
+      let submitProgress = 0;
+      let progressInterval: NodeJS.Timeout | undefined = undefined;
+
+      if (hasFiles(values)) {
+        progressInterval = setInterval(() => {
+          if (submitProgress <= 0.1) {
+            submitProgress += 0.02;
+            document.dispatchEvent(
+              new CustomEvent(EventKeys.submitProgress, {
+                detail: {
+                  progress: submitProgress,
+                  message: formikBag.props.t("submitProgress.text"),
+                },
+              })
+            );
+          } else {
+            clearInterval(progressInterval);
+          }
+        }, 500);
+      }
+
       const result = await submitForm(
         formValuesWithoutFileContent,
         formikBag.props.language,
         formikBag.props.formRecord.id,
         formikBag.props.captchaToken?.current
       );
+
+      clearInterval(progressInterval);
 
       // Failed to find Server Action (likely due to newer deployment)
       if (result === undefined) {
@@ -368,6 +393,11 @@ export const Form = withFormik<FormProps, Responses>({
               const totalProgress =
                 Object.values(fileProgress).reduce((acc, progress) => acc + progress, 0) /
                 totalFiles;
+
+              if (totalProgress <= submitProgress) {
+                // Don't dispatch progress events if the total progress is less than what we've already dispatched
+                return;
+              }
 
               document.dispatchEvent(
                 new CustomEvent(EventKeys.submitProgress, {