+++ packages/types/src/form-types.ts
@@ -1,3 +1,5 @@
+import { type FileInputResponse } from "./form-response-types";
+
 // Utility type creator
 export type TypeOmit<Type, Key extends PropertyKey> = {
   [Property in keyof Type as Exclude<Property, Key>]: Type[Property];
@@ -265,3 +267,23 @@ export const NotificationsIntervalDefault = NotificationsInterval.DAY;
 
 export type NotificationsInterval =
   (typeof NotificationsInterval)[keyof typeof NotificationsInterval];
+
+export type DateFormat = "YYYY-MM-DD" | "DD-MM-YYYY" | "MM-DD-YYYY";
+
+export interface DateObject {
+  YYYY: number;
+  MM: number;
+  DD: number;
+}
+
+export enum DatePart {
+  DD = "day",
+  MM = "month",
+  YYYY = "year",
+}
+
+export interface FileInput extends FileInputResponse {
+  name: string;
+  size: number;
+  content: ArrayBuffer;
+}