+++ packages/types/src/index.ts
@@ -22,6 +22,12 @@ import {
   NotificationsInterval,
   NotificationsIntervalDefault,
   BetaFormElementTypes,
+  DateFormat,
+  DateObject,
+  DatePart,
+  GroupsType,
+  FormValues,
+  FileInput,
 } from "./form-types";
 
 export type { Response, Responses, FileInputResponse } from "./form-response-types";
@@ -47,6 +53,29 @@ export { FormStatus };
 export { NotificationsInterval };
 export { NotificationsIntervalDefault };
 export { BetaFormElementTypes };
+export type { DateFormat, DateObject, DatePart };
+export type { GroupsType };
+export type { FormValues };
 
 export { SortOption };
 export type { SortValue };
+
+import {
+  AddressCompleteChoice,
+  AddressCompleNext,
+  AddressCompleteResult,
+  AddressCompleteProps,
+  AddressElements,
+  AddressCompleteLabels,
+} from "./address-complete";
+
+export type {
+  AddressCompleteChoice,
+  AddressCompleNext,
+  AddressCompleteResult,
+  AddressCompleteProps,
+  AddressElements,
+  AddressCompleteLabels,
+};
+
+export type { FileInput };