+++ packages/types/CHANGELOG.md
@@ -5,9 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.15] - 2025-08-15
+## [1.0.18] - 2025-08-165
 ### Changed
 - Add FormBetaElementType
+
+## [1.0.17] - 2025-08-15
+
+- Add FileInput type
+
+## [1.0.16] - 2025-08-13
+
+- Add DateObject to Response definition
+
+## [1.0.15] - 2025-08-13
+
+### Changed
+
+- Update fileInputResponse to migrate from base64 to BufferArray
 
 ## [1.0.14] - 2025-07-10